requirements.txt
# Web scraping libraries
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2

# Data handling
pandas==2.1.3
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
fake-useragent==1.4.0

# Logging
loguru==0.7.2

# Email notifications (optional)
python-decouple==3.8
